add oteltest package
