+++ x-pack/metricbeat/mbreceiver/receiver_test.go
@@ -5,13 +5,12 @@
 package mbreceiver
 
 import (
-	"bytes"
 	"context"
 	"sync/atomic"
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/assert"
+	"github.com/elastic/beats/v7/libbeat/otelbeat/oteltest"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/pdata/plog"
@@ -91,89 +90,40 @@ found:
 }
 
 func TestMultipleReceivers(t *testing.T) {
-	logs := make(map[string]int)
-
-	ctx := context.Background()
-	createReceiver := func(t *testing.T, name string) receiver.Logs {
-		t.Helper()
-		config := Config{
-			Beatconfig: map[string]interface{}{
-				"metricbeat": map[string]interface{}{
-					"modules": []map[string]interface{}{
-						{
-							"module":     "system",
-							"enabled":    true,
-							"period":     "1s",
-							"processes":  []string{".*"},
-							"metricsets": []string{"cpu"},
-						},
+	config := Config{
+		Beatconfig: map[string]interface{}{
+			"metricbeat": map[string]interface{}{
+				"modules": []map[string]interface{}{
+					{
+						"module":     "system",
+						"enabled":    true,
+						"period":     "1s",
+						"processes":  []string{".*"},
+						"metricsets": []string{"cpu"},
 					},
 				},
-				"output": map[string]interface{}{
-					"otelconsumer": map[string]interface{}{},
-				},
-				"logging": map[string]interface{}{
-					"level": "debug",
-					"selectors": []string{
-						"*",
-					},
-				},
-				"path.home": t.TempDir(),
 			},
-		}
-
-		var zapLogs bytes.Buffer
-		core := zapcore.NewCore(
-			zapcore.NewJSONEncoder(zap.NewProductionEncoderConfig()),
-			zapcore.AddSync(&zapLogs),
-			zapcore.DebugLevel)
-
-		receiverSettings := receiver.Settings{}
-		receiverSettings.Logger = zap.New(core).Named(name)
-
-		logConsumer, err := consumer.NewLogs(func(ctx context.Context, ld plog.Logs) error {
-			for i := 0; i < ld.ResourceLogs().Len(); i++ {
-				rl := ld.ResourceLogs().At(i)
-				for j := 0; j < rl.ScopeLogs().Len(); j++ {
-					sl := rl.ScopeLogs().At(j)
-					for k := 0; k < sl.LogRecords().Len(); k++ {
-						log := sl.LogRecords().At(k)
-						logs[name] = logs[name] + 1
-						t.Logf("ingested log for %q: %v", name, log.Body().Map().AsRaw())
-					}
-				}
-			}
-			return nil
-		})
-		assert.NoErrorf(t, err, "Error creating log consumer for %q", name)
-
-		t.Cleanup(func() {
-			if t.Failed() {
-				t.Logf("Logs for %q: %s\n", name, zapLogs.String())
-			}
-		})
-
-		r, err := NewFactory().CreateLogs(ctx, receiverSettings, &config, logConsumer)
-		assert.NoErrorf(t, err, "Error creating receiver %q", name)
-		return r
+			"output": map[string]interface{}{
+				"otelconsumer": map[string]interface{}{},
+			},
+			"logging": map[string]interface{}{
+				"level": "debug",
+				"selectors": []string{
+					"*",
+				},
+			},
+			"path.home": t.TempDir(),
+		},
 	}
 
-	r1 := createReceiver(t, "r1")
-	r2 := createReceiver(t, "r2")
-
-	err := r1.Start(ctx, nil)
-	require.NoError(t, err, "Error starting receiver 1")
-	defer func() {
-		require.NoError(t, r1.Shutdown(ctx))
-	}()
-
-	err = r2.Start(ctx, nil)
-	require.NoError(t, err, "Error starting receiver 2")
-	defer func() {
-		require.NoError(t, r2.Shutdown(ctx))
-	}()
-
-	require.Eventuallyf(t, func() bool {
-		return logs["r1"] > 1 && logs["r2"] > 1
-	}, 1*time.Minute, 100*time.Millisecond, "timeout waiting for logs: %#v", logs)
+	oteltest.CheckMultipleReceivers(oteltest.CheckMultipleReceiversParams{
+		T:       t,
+		Factory: NewFactory(),
+		Config:  &config,
+		AssertFunc: func(t *testing.T, logs map[string]int) {
+			require.Eventuallyf(t, func() bool {
+				return logs["r1"] > 1 && logs["r2"] > 1
+			}, 1*time.Minute, 100*time.Millisecond, "timeout waiting for logs: %#v", logs)
+		},
+	})
 }
